Moving preferences files to correct directory.
