+++ shared/src/desktopMain/kotlin/template/shared/data/preferences/CreateDataStore.desktop.kt
@@ -1,4 +1,4 @@
-package template.shared
+package template.shared.data.preferences
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
